requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
cloudinary==1.44.1
crispy-bootstrap5==0.7
Cython==3.1.2
dj-database-url==3.0.1
Django==5.1.3
django-autocomplete-light==3.12.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
